requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-appconf==1.0.5
django-cors-headers==3.11.0
django-imagekit==4.1.0
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
flake8==4.0.1
fontawesome-free==5.15.4
mccabe==0.6.1
phonenumberslite==8.12.44
pilkit==2.0
Pillow==9.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
